Updated examples to v2 using Leptos.
