--- examples/leptos/.gitignore
@@ -1,3 +1,0 @@
-/dist/
-/target/
-/Cargo.lock
