--- examples/leptos/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "test-tauri-events-ui"
-version = "0.0.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dependencies]
-leptos = { version = "0.6", features = ["csr", "nightly"] }
-wasm-bindgen = "0.2"
-wasm-bindgen-futures = "0.4"
-js-sys = "0.3"
-serde = { version = "1", features = ["derive"] }
-serde-wasm-bindgen = "0.6"
-console_error_panic_hook = "0.1.7"
-tauri-sys = { path = "../tauri-sys", features = ["all"] }
-futures = "0.3"
-tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
-tracing-web = "0.1.3"
-
-[workspace]
-members = ["src-tauri"]
