--- examples/leptos/Trunk.toml
@@ -1,10 +1,0 @@
-[build]
-target = "./index.html"
-
-[watch]
-ignore = ["./src-tauri"]
-
-[serve]
-address = "127.0.0.1"
-port = 1420
-open = false
