--- examples/leptos/src-tauri/.gitignore
@@ -1,7 +1,0 @@
-# Generated by Cargo
-# will have compiled files and executables
-/target/
-
-# Generated by Tauri
-# will have schema files for capabilities auto-completion
-/gen/schemas
