--- examples/leptos/src-tauri/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "test-tauri-events"
-version = "0.0.0"
-description = "A Tauri App"
-authors = ["you"]
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[build-dependencies]
-# tauri-build = { version = "2.0.0-beta", features = [] }
-tauri-build = { path = "../../tauri/core/tauri-build", features = [] }
-
-[dependencies]
-tauri = { path = "../../tauri/core/tauri", features = [] }
-tauri-plugin-shell = { path = "../../plugins-workspace/plugins/shell" }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
