--- examples/leptos/src-tauri/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    tauri_build::build()
-}
