+++ examples/test/.vscode/extensions.json
