+++ examples/test/Cargo.toml
@@ -1,19 +1,19 @@
 [package]
-name = "tauri-sys-test-ui"
+name = "test-tauri-events-ui"
 version = "0.0.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
-tauri-sys = { path = "../../", features = ["all"] }
-sycamore = { git = "https://github.com/sycamore-rs/sycamore", rev = "abd556cbc02047042dad2ebd04405e455a9b11b2", features = ["suspense"] }
-anyhow = "1.0.75"
+leptos = { version = "0.6", features = ["csr", "nightly"] }
+wasm-bindgen = "0.2"
+wasm-bindgen-futures = "0.4"
+js-sys = "0.3"
+serde = { version = "1", features = ["derive"] }
+serde-wasm-bindgen = "0.6"
 console_error_panic_hook = "0.1.7"
-wasm-bindgen-futures = "0.4.39"
-serde = { version = "1.0.193", features = ["derive"] }
-log = { version = "0.4.20", features = ["serde"] }
-futures = "0.3.29"
-gloo-timers = { version = "0.3", features = ["futures"] }
-
-[features]
-ci = []
+tauri-sys = { path = "../../", features = ["all"] }
+futures = "0.3"
+tracing = "0.1.40"
+tracing-subscriber = "0.3.18"
+tracing-web = "0.1.3"
