+++ examples/test/index.html
@@ -1,8 +1,11 @@
-<!DOCTYPE html>
+<!doctype html>
 <html>
   <head>
     <meta charset="utf-8" />
-    <title>Tauri + Yew App</title>
-    <link data-trunk rel="css" href="./styles.css">
+    <title>Tauri + Leptos App</title>
+    <link data-trunk rel="css" href="styles.css" />
+    <link data-trunk rel="copy-dir" href="public" />
+    <link data-trunk rel="rust" data-wasm-opt="z" />
   </head>
+  <body></body>
 </html>
