+++ examples/test/src-tauri/.gitignore
@@ -2,3 +2,6 @@
 # will have compiled files and executables
 /target/
 
+# Generated by Tauri
+# will have schema files for capabilities auto-completion
+/gen/schemas
