+++ examples/test/src-tauri/capabilities/default.json
