--- examples/test/src-tauri/ci.tauri.conf.json
@@ -1,6 +1,0 @@
-{
-  "build": {
-    "beforeDevCommand": "trunk serve --features ci",
-    "beforeBuildCommand": "trunk build --release --features ci"
-  }
-}
