+++ examples/test/src-tauri/src/main.rs
@@ -1,70 +1,41 @@
-#![cfg_attr(
-    all(not(debug_assertions), target_os = "windows"),
-    windows_subsystem = "windows"
-)]
+// Prevents additional console window on Windows in release, DO NOT REMOVE!!
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
-use std::sync::atomic::{AtomicBool, Ordering};
-use tauri::{Manager, Runtime, State, Window};
-use tauri_plugin_log::{LogTarget, LoggerBuilder};
-
-struct Received(AtomicBool);
-// Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
-#[tauri::command]
-fn verify_receive(emitted: State<Received>) -> bool {
-    emitted.0.load(Ordering::Relaxed)
-}
-
-#[tauri::command]
-async fn emit_event<R: Runtime>(win: Window<R>) -> Result<(), ()> {
-    let _ = win.emit("rust-event-once", "Hello World from Rust!");
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn emit_event_5_times<R: Runtime>(win: Window<R>) -> Result<(), ()> {
-    for i in 0..5 {
-        let _ = win.emit("rust-event-listen", i);
-    }
-
-    Ok(())
-}
-
-#[tauri::command]
-fn exit_with_error(e: &str) -> bool {
-    eprintln!("{}", e);
-    std::process::exit(1);
-}
+use tauri::{Emitter, Manager};
 
 fn main() {
-    let log_plugin = {
-        let targets = [
-            LogTarget::LogDir,
-            #[cfg(debug_assertions)]
-            LogTarget::Stdout,
-            #[cfg(debug_assertions)]
-            LogTarget::Webview,
-        ];
-
-        LoggerBuilder::new().targets(targets).build()
-    };
-
+    logging::enable();
     tauri::Builder::default()
-        .plugin(log_plugin)
-        .invoke_handler(tauri::generate_handler![verify_receive, emit_event, emit_event_5_times, exit_with_error])
-        .setup(|app| {
-            app.manage(Received(AtomicBool::new(false)));
-
-            let app_handle = app.handle();
-            app.listen_global("javascript-event", move |_| {
-                app_handle
-                    .state::<Received>()
-                    .0
-                    .store(true, Ordering::Relaxed);
-            });
-
-            Ok(())
-        })
+        .invoke_handler(tauri::generate_handler![trigger_listen_events,])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
+}
+
+#[tauri::command]
+fn trigger_listen_events(app: tauri::AppHandle) {
+    tracing::debug!("trigger_listen_event");
+    std::thread::spawn({
+        move || {
+            for i in 1..=100 {
+                app.emit("event::listen", i).unwrap();
+                std::thread::sleep(std::time::Duration::from_millis(500));
+            }
+        }
+    });
+}
+
+mod logging {
+    use tracing_subscriber::{filter::LevelFilter, fmt, prelude::*, Layer, Registry};
+
+    const MAX_LOG_LEVEL: LevelFilter = LevelFilter::DEBUG;
+
+    pub fn enable() {
+        let console_logger = fmt::layer()
+            .with_writer(std::io::stdout)
+            .pretty()
+            .with_filter(MAX_LOG_LEVEL);
+
+        let subscriber = Registry::default().with(console_logger);
+        tracing::subscriber::set_global_default(subscriber).unwrap();
+    }
 }
