--- examples/test/src/clipboard.rs
@@ -1,12 +1,0 @@
-use anyhow::ensure;
-use tauri_sys::clipboard;
-
-pub async fn test() -> anyhow::Result<()> {
-    clipboard::write_text("foobar").await?;
-
-    let text = clipboard::read_text().await?;
-
-    ensure!(text == "foobar".to_string());
-
-    Ok(())
-}
